with bug on end form

+++ Acars/FlightData/FlightDatabase.cs
@@ -178,10 +178,10 @@ namespace Acars.FlightData
                 MySqlCommand sqlCmd = new MySqlCommand(sqlStrInsertPirep, conn);
                 var dateParam = sqlCmd.Parameters.Add("@date", MySqlDbType.Date);
                 dateParam.Value = DateTime.UtcNow;
-                sqlCmd.Parameters.AddWithValue("@flighttime", Math.Round(flight.ActualTimeEnRoute.TotalMinutes));
+                sqlCmd.Parameters.AddWithValue("@flighttime", (int)Math.Round(flight.ActualTimeEnRoute.TotalMinutes));
                 sqlCmd.Parameters.AddWithValue("@flightid", flight.FlightID);
                 sqlCmd.Parameters.AddWithValue("@pilotid", Properties.Settings.Default.Email);
-                sqlCmd.Parameters.AddWithValue("@landingrate", Math.Round(flight.ActualArrivalTime.VerticalSpeed));
+                sqlCmd.Parameters.AddWithValue("@landingrate", (int)Math.Round(flight.ActualArrivalTime.VerticalSpeed));
                 sqlCmd.Parameters.AddWithValue("@sum", flight.FinalScore);
                 sqlCmd.Parameters.AddWithValue("@accepted", "1");
                 sqlCmd.Parameters.AddWithValue("@flighteps", flight.EfficiencyPoints);
